Split test for loading bcf and excel
