+++ tests/cli/load/test_cli_load.py
@@ -1,4 +1,5 @@
 """Test the CLI load functions"""
+
 import logging
 from pathlib import Path
 
@@ -37,6 +38,25 @@ def test_load_bcf_analysis(
     assert Sample.query.get(vcf_sample_id).analyses[0] == new_analysis
 
 
+def test_load_bcf_analysis_sex(
+    cli_runner: CliRunner, bcf_path: Path, snp_db: Manager, vcf_sample_id: str, caplog
+):
+    """Test to load a sequence analysis and check if there was any sex added"""
+    # GIVEN a database with some SNPs loaded and one sample
+    # GIVEN a vcf with one sample
+
+    context = {"db": snp_db}
+    # WHEN loading a BCF from the command line (single sample)
+    result = cli_runner.invoke(load_cmd, [str(bcf_path)], obj=context)
+
+    sample_obj = Sample.query.first()
+    # THEN assert that the sample has no sex since there is no such information
+    assert sample_obj.sex is None
+    # THEN assert that the analysis have no sex
+    analysis_obj = Analysis.query.first()
+    assert analysis_obj.sex is None
+
+
 def test_load_bcf_analysis_exists(
     cli_runner: CliRunner, bcf_path: Path, sequence_db: Manager, vcf_sample_id: str, caplog
 ):
@@ -54,45 +74,6 @@ def test_load_bcf_analysis_exists(
     # THEN it should exit succesfully but log a warning
     assert result.exit_code == 0
     assert "found previous analysis, skip" in caplog.text
-
-
-def test_load_genotype_analysis(cli_runner: CliRunner, excel_single_path: Path, snp_db: Manager):
-    """Test to load a genotype analysis from a excel document"""
-    # GIVEN a database with some SNPs loaded and no samples
-    assert SNP.query.count() > 0
-    assert Sample.query.count() == 0
-    context = {"db": snp_db}
-
-    # WHEN loading an Excel book from the command line (multi-sample)
-    result = cli_runner.invoke(load_cmd, [str(excel_single_path), "-k", "ID-CG-"], obj=context)
-
-    # THEN it works and loads a sample with an analysis
-    assert result.exit_code == 0
-    assert Sample.query.count() > 0
-    analysis_obj = Analysis.query.first()
-    assert analysis_obj.type == "genotype"
-
-
-def test_load_genotype_analysis_existing_sequence(
-    cli_runner: CliRunner, excel_single_path: Path, sequence_db: Manager
-):
-    """Test to load a genotype analysis from a excel document when the sequence analysis exists"""
-
-    # GIVEN a database with some SNPs loaded and one sample
-    assert SNP.query.count() > 0
-    assert Sample.query.count() == 1
-    assert Analysis.query.count() == 1
-    context = {"db": sequence_db}
-
-    # WHEN loading an Excel book from the command line (multi-sample)
-    result = cli_runner.invoke(load_cmd, [str(excel_single_path), "-k", "ID-CG-"], obj=context)
-
-    # THEN it works and loads a sample with an analysis
-    assert result.exit_code == 0
-    # THEN assert that there is still one sample in the database
-    assert Sample.query.count() == 1
-    # THEN assert that two analysies exist
-    assert Analysis.query.count() == 2
 
 
 def test_load_unknown_format(cli_runner: CliRunner, config_path: Path, snp_db: Manager, caplog):
